restructure token/api
