+++ src/main/java/com/thalesdemo/safenet/server/controller/HealthController.java
@@ -46,8 +46,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.thalesdemo.safenet.auth.api.Authenticate;
-import com.thalesdemo.safenet.token.api.ApiException;
-import com.thalesdemo.safenet.token.api.PingService;
+import com.thalesdemo.safenet.token.api.exception.ApiException;
+import com.thalesdemo.safenet.token.api.service.PingService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
