+++ src/main/java/com/thalesdemo/safenet/token/api/controller/ApiController.java
@@ -1,4 +1,4 @@
-package com.thalesdemo.safenet.token.api;
+package com.thalesdemo.safenet.token.api.controller;
 // package com.thalesdemo.safenet.token.list.api;
 
 // import java.io.PrintWriter;
