+++ src/main/java/com/thalesdemo/safenet/token/api/dto/GetTotalTokensResponseDTO.java
@@ -1,4 +1,4 @@
-package com.thalesdemo.safenet.token.api;
+package com.thalesdemo.safenet.token.api.dto;
 
 import lombok.Getter;
 import lombok.Setter;
