+++ src/main/java/com/thalesdemo/safenet/token/api/exception/ApiException.java
@@ -1,4 +1,4 @@
-package com.thalesdemo.safenet.token.api;
+package com.thalesdemo.safenet.token.api.exception;
 
 public class ApiException extends RuntimeException {
     public ApiException(String message) {
