+++ src/main/java/com/thalesdemo/safenet/token/api/exception/ApiExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.thalesdemo.safenet.token.api;
+package com.thalesdemo.safenet.token.api.exception;
 
 import java.util.HashMap;
 import java.util.Map;
