+++ src/main/java/com/thalesdemo/safenet/token/api/request/GetTokenSoapRequest.java
@@ -1,4 +1,4 @@
-package com.thalesdemo.safenet.token.api.requests;
+package com.thalesdemo.safenet.token.api.request;
 
 import javax.xml.soap.MessageFactory;
 import javax.xml.soap.SOAPBody;
@@ -9,33 +9,34 @@ import javax.xml.soap.SOAPException;
 import javax.xml.soap.SOAPMessage;
 import javax.xml.soap.SOAPPart;
 
-public class GetUserSoapRequest {
+public class GetTokenSoapRequest {
 
-    private GetUserSoapRequest() {
+    private GetTokenSoapRequest() {
         throw new IllegalStateException("Utility class");
     }
 
-    public static SOAPMessage createGetUserRequest(String userName, String organization) throws SOAPException {
+    public static SOAPMessage createGetTokenRequest(String serial, String organization) throws SOAPException {
         MessageFactory factory = MessageFactory.newInstance(SOAPConstants.SOAP_1_2_PROTOCOL);
         SOAPMessage message = factory.createMessage();
         SOAPPart soapPart = message.getSOAPPart();
         SOAPEnvelope envelope = soapPart.getEnvelope();
+
         envelope.addNamespaceDeclaration("xsi", "http://www.w3.org/2001/XMLSchema-instance");
         envelope.addNamespaceDeclaration("xsd", "http://www.w3.org/2001/XMLSchema");
         envelope.addNamespaceDeclaration("soap12", "http://www.w3.org/2003/05/soap-envelope");
 
         SOAPBody body = envelope.getBody();
 
-        // Create the "GetUser" element and its child elements
-        SOAPElement getUserElement = body.addChildElement("GetUser", "", "http://www.cryptocard.com/blackshield/");
+        // Create the "GetToken" element and its child elements
+        SOAPElement getTokenElement = body.addChildElement("GetToken", "", "http://www.cryptocard.com/blackshield/");
 
-        if (userName != null) {
-            SOAPElement userNameElement = getUserElement.addChildElement("userName");
-            userNameElement.addTextNode(userName);
+        if (serial != null) {
+            SOAPElement serialElement = getTokenElement.addChildElement("serial");
+            serialElement.addTextNode(serial);
         }
 
         if (organization != null) {
-            SOAPElement organizationElement = getUserElement.addChildElement("organization");
+            SOAPElement organizationElement = getTokenElement.addChildElement("organization");
             organizationElement.addTextNode(organization);
         }
