+++ src/main/java/com/thalesdemo/safenet/token/api/service/PingService.java
@@ -1,9 +1,11 @@
-package com.thalesdemo.safenet.token.api;
+package com.thalesdemo.safenet.token.api.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import com.thalesdemo.safenet.token.api.exception.ApiException;
 
 @Service
 public class PingService {
